+++ mocks/browser.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw';
-import { handlers } from './handler';
+import { handlers } from './handler/handler';
 
 export const worker = setupWorker(...handlers);
