+++ mocks/handler/dashboard.ts/mission.ts
@@ -1,11 +1,11 @@
 import { rest } from 'msw';
-import Config from '../config/config.export';
-import { Missions } from './data/Missions';
-import { API_URL_DASHBOARD } from '../pages/api/dashboard';
+import Config from '../../../config/config.export';
+import { Missions } from '../../data/Missions';
+import { API_URL_DASHBOARD } from '../../../pages/api/dashboard';
 
 const missions = Missions();
 
-export const dashboardMissionHandler = [
+export const missionHandler = [
   rest.get(
     `${Config().baseUrl}${API_URL_DASHBOARD.MISSION}`,
     async (req, res, ctx) => {
