+++ mocks/handler/handler.ts
@@ -1,0 +1,9 @@
+import { otherHandler } from './other';
+import { lectureHandler } from './lecture/lecture';
+import { dashboardHandler } from './dashboard.ts/dashboard';
+
+export const handlers = [
+  ...lectureHandler,
+  ...dashboardHandler,
+  ...otherHandler,
+];
