+++ mocks/handler/lecture/lecture.ts
@@ -1,0 +1,15 @@
+import { rest } from 'msw';
+import Config from '../../../config/config.export';
+import { lectureroomsData } from '../../data/lectureroomsData';
+import { commentHandler } from './comment';
+import { API_URL_LECTURE } from '../../../pages/api/lecture';
+
+export const lectureHandler = [
+  rest.get(
+    `${Config().baseUrl}${API_URL_LECTURE.ALL_LECTURE}`,
+    async (req, res, ctx) => {
+      return res(ctx.json(lectureroomsData));
+    },
+  ),
+  ...commentHandler,
+];
