+++ mocks/handler/other.ts
@@ -1,7 +1,7 @@
 import { rest } from 'msw';
-import Config from '../config/config.export';
-import { API_URL_OTHER } from '../pages/api/other';
-import { calendarData } from './data/calendarData';
+import Config from '../../config/config.export';
+import { calendarData } from '../data/calendarData';
+import { API_URL_OTHER } from '../../pages/api/other';
 
 export const otherHandler = [
   rest.get(
