+++ mocks/server.ts
@@ -1,4 +1,4 @@
 import { setupServer } from 'msw/node';
-import { handlers } from './handler';
+import { handlers } from './handler/handler';
 
 export const server = setupServer(...handlers);
